added get request for aboutMe data, created About List Component and State for aboutMe
