+++ nextjs/components/about/AboutParagraph.jsx
@@ -3,7 +3,7 @@ export default function AboutParagraph({ data }){
   return (
     <div className="relative">
       <p className="my-4 text-center  sm:text-lg md:text-xl">
-        { data }
+        { data.text }
       </p>
     </div>
   );
